Improve code by refactoring and adding new features
This commit includes a refactoring of the code and addition of new features. Specifically, we have marked the `IdentityParser` class as internal for added privacy, created a new class `IdentityStringGenerator` to generate unique identifiers for interface and name combinations, and developed a new test for the `IdentityGenerator` class. We also refactored the `Psr11InjectorTest.php` to utilise the new `IdentityGenerator` class. These changes enhance readability and maintainability of the code.